Abstract defaultLayouts for all MDX
