+++ examples/kitchen-sink/content/non-page-default-layout.mdx
@@ -1,0 +1,1 @@
+Some content
