+++ packages/gatsby-mdx/utils/gen-mdx.js
@@ -1,13 +1,86 @@
+const _ = require("lodash");
 const babel = require("@babel/core");
 const grayMatter = require("gray-matter");
 const mdx = require("@mdx-js/mdx");
 const objRestSpread = require("@babel/plugin-proposal-object-rest-spread");
+
+const slash = require("slash");
+const unified = require("unified");
+const squeeze = require("remark-squeeze-paragraphs");
+const toMDAST = require("remark-parse");
+const {
+  isImport,
+  isExport,
+  isExportDefault,
+  BLOCKS_REGEX,
+  EMPTY_NEWLINE
+} = require("@mdx-js/mdx/util");
 
 const debug = require("debug")("gatsby-mdx:gen-mdx");
 
 const getSourcePluginsAsRemarkPlugins = require("./get-source-plugins-as-remark-plugins");
 const htmlAttrToJSXAttr = require("./babel-plugin-html-attr-to-jsx-attr");
 const BabelPluginPluckImports = require("./babel-plugin-pluck-imports");
+
+const DEFAULT_OPTIONS = {
+  footnotes: true,
+  remarkPlugins: [],
+  rehypePlugins: [],
+  compilers: [],
+  blocks: [BLOCKS_REGEX]
+};
+
+/**
+ * TODO: Find a way to PR all of this code that was lifted
+ * from @mdx-js/mdx back into mdx with the modifications. We
+ * don't want to maintain subtly different parsing code if we
+ * can avoid it.
+ */
+const hasDefaultExport = (str, options) => {
+  let hasDefaultExportBool = false;
+
+  function getDefaultExportBlock(subvalue) {
+    const isDefault = isExportDefault(subvalue);
+    hasDefaultExportBool = hasDefaultExportBool || isDefault;
+    return isDefault;
+  }
+  const tokenizeEsSyntax = (eat, value) => {
+    const index = value.indexOf(EMPTY_NEWLINE);
+    const subvalue = value.slice(0, index);
+
+    if (isExport(subvalue) || isImport(subvalue)) {
+      return eat(subvalue)({
+        type: isExport(subvalue) ? "export" : "import",
+        default: getDefaultExportBlock(subvalue),
+        value: subvalue
+      });
+    }
+  };
+
+  tokenizeEsSyntax.locator = value => {
+    return isExport(value) || isImport(value) ? -1 : 1;
+  };
+
+  function esSyntax() {
+    var Parser = this.Parser;
+    var tokenizers = Parser.prototype.blockTokenizers;
+    var methods = Parser.prototype.blockMethods;
+
+    tokenizers.esSyntax = tokenizeEsSyntax;
+
+    methods.splice(methods.indexOf("paragraph"), 0, "esSyntax");
+  }
+
+  const { content } = grayMatter(str);
+  unified()
+    .use(toMDAST, options)
+    .use(esSyntax)
+    .use(squeeze, options)
+    .parse(content)
+    .toString();
+
+  return hasDefaultExportBool;
+};
 
 /*
  * function mutateNode({
@@ -114,8 +187,36 @@ export const _frontmatter = ${JSON.stringify(data)}`;
     }
   );
 
+  const parent = node.parent && getNode(node.parent);
+  const source =
+    parent && parent.internal.type === `File` && parent.sourceInstanceName;
+
+  // get the default layout for the file source group, or if it doesn't
+  // exist, the overall default layout
+  const defaultLayout = _.get(
+    options.defaultLayouts,
+    source,
+    _.get(options.defaultLayouts, "default")
+  );
+
+  let code = content;
+  // after running mdx, the code *always* has a default export, so this
+  // check needs to happen first.
+  if (!hasDefaultExport(content, DEFAULT_OPTIONS) && !!defaultLayout) {
+    debug("inserting default layout", defaultLayout);
+    const { content: contentWithoutFrontmatter, matter } = grayMatter(content);
+
+    code = `${matter ? matter : ""}
+
+import DefaultLayout from "${slash(defaultLayout)}"
+
+export default DefaultLayout
+
+${contentWithoutFrontmatter}`;
+  }
+
   debug("running mdx");
-  let code = await mdx(content, {
+  code = await mdx(code, {
     ...options,
     remarkPlugins: options.remarkPlugins.concat(
       gatsbyRemarkPluginsAsremarkPlugins
